requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-genai>=0.8.0
groq>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
